sticking to the maze
I decided my maze is intricate enough to hopefully mean something. it's already making me struggle as it is. so I generated a maze and then drew it out manually and then struggled to make my previous code work with it. so far I think it works fine! :)

the plan is to add hidden doors and keys and turn it into a puzzle. for now I'm struggling with commands. the speech recognition reads me saying numbers as a string: '2' as 'two' etc. so it just doesn't move the way I want it to... not sure how to fix this