fixed!!!!!!!!!
and it works!!!!! I added an object of written out numbers to make up for the fact that sometimes it the voice recognizer writes them out (ie. '2' as 'two').

I added a transcript at the bottom to write out what the voice says. I tweaked the voice to go faster but when I select different voices than the default they sometimes randomly break

I also just now discovered that NaN and null do not mean the same thing. after all my years of coding. YAAAAAY

anyway the voice recognition stuff works all that's left is to add the puzzle items and interactions