actually almost done
Pippin helped me sort a randomizer for the key and door colors. I made the maze invisible and begin displaying when you ask the player to check whats in front. I also added the exit and win condition works. All that's left to do in the title and win screens :)