hi.
this is my first commit despite working on this for 3 days because I could just NOT figure out phaser!!! and it took me this long to even begin to understand how to execute my idea and make it work. everything from now on is going to be really trial and error. for now, I just want to get the card selection done and I'll figure out how to do the rest when I get to ask Pippin next week...