oh my god
its 1am but I FINALLY made the stupid scene change work because for some reason Battle just kept giving me the same error over and over (Scene not found for key: battle). I switched Battle to be before Play in the main array and suddenly it worked. This thing sucks