started battle
added all enemies to the game and setup their displays, made the animation when battle starts and setup the necessary functions for the battle to play out. I just need to add the animations for when the battle plays out.

on a better note, I finally understand tweens OnComplete thanks to Pippin and they work a lot better for the rest of my game. it also took me a while to figure this out, but I need to remove event listeners I add to each object depending on the scene (with on and off)