battle done (?)
finished the battle! added animations and everything. I also fixed the card stacking animation so that it's always consistent. I started setting up the shop too. the only things left to do for the battle is the death and swap, but to do those I need to actually be able to test it with multiple cards in the deck and right now I can't because the battle just randomly breaks on second round and I hate javascript. I asked Pippin for help and hopefully I'll figure this out by tomorrow...