oh my god
IM SOOOOOOO STUPID turns out I didnt realize I had a *second* this.battleMenu that I changed to a string at the end of the code there. because I am stupid and named the variable the same thing. which is why it no longer knew what to look for. Ok well everything works now so Ill move on to making the swap and then the shop menu