upgrade finished and added saki ability
I'm going to play through the game a few times first now that most basics are implemented and see if there's any fixes to be made before adding rui, kasa and emu special abilities.

I also need to add card deaths from exhaustion/hunger. they shouldn't take too much time but the timing of when it and the abilities should be implemented is a bit of a headache for me. I also don't know if I should make it a separate scene or not?? it's a lot and it has its own separate events I guess...?