urgh
scrapping additional skills because Im just out of time. I got worn out working on this for so long and my game is fairly intricate as is. I noted down what the scrapped skills are and why I won't have time to make them. I'll spend the rest of my time re-balancing the game because it's pretty hard to beat right now!!